Update import paths to use alias @ and move types from login and register to @/types
